Make it possible to build as header only library
